TEST: Select db provider from value of env var TEKUMA_TEST_DB
If the environment variable TEKUMA_TEST_DB is not defined or has the
value 'sqlite' (or some variant spelling, e.g., 'SQLite' or
'sqlite3'), then the database software used is SQLite.

If TEKUMA_TEST_DB = 'mysql' (or some variant spelling, e.g., 'MySQL'),
then a MySQL database is used for testing. Note that in this case, a
server must be available, and the configuration tests/testdbconf.json
will be used.