DOC: Delete copy of schema from comments of SearchMain.jsx
There is a danger that the copy in the source code comments drifts
away from the actual specification, e.g., because it is not kept
up-to-date. Thus, we should try to keep only one main point of storage
for the schema. If needed, direct readers to there instead of making
local copies.