Create dedicated local db config for manual tests
This configuration is for manual testing, without using `npm run
testserv`. E.g., while developing, it is useful to maintain a local
database that is similar to the production database and that has
persistent state, whereas the full test suite needs to setup and tear
down the entire database, possibly more than once.

Now the full test suite, when using a MySQL server, connects to port 3307.